Split sidebar scheduler implementation
